SW-2815: audio: apbridgea: Register CPorts for specific directions
Currently, it is assumed that all audio data CPorts registered on
APB1 are used for transmitting audio data.  That may not always be
true like when a microphone is connected but no speakers.  Also,
the current protocol lacks a way to tell APB1 whether the CPort
being registered is for transmitting, receiving, or both.

Fix by adding a 'direction' field to the register and unregister CPort
requests and define bits indicating which direction (or both) audio
data will go on that CPort.

Testing Done: Played audio on an EVT1.5 + EVT1 & EVT1.5 audio modules

Change-Id: I31e92ca7d39daf0a7ab91b13dc2cd85851906018
Signed-off-by: Mark Greer <mgreer@animalcreek.com>