SW-2815: audio: apbridgea: Add receive support
Add support for receiving an audio stream from a single UniPro
audio module.  For efficiency, cache the cport ID of the connection
the data is being received on in the info structure instead of
searching through the cport list.

Testing Done: Played audio on an EVT1.5 + EVT1 & EVT1.5 audio modules.
              Cannot test receiving because necessary module support is
              not there yet.

Change-Id: Iff5ba2d0b3f26c7bd9d198ff5f21218d63f33241
Signed-off-by: Mark Greer <mgreer@animalcreek.com>