SW-2085: SVC: ES2 session data read: return actual data length
The Switch can signal different data lengths in the status
report and in the actual session data header. This is the case
when data reception is busy or when multiple incoming messages
have been received.
Retrieve the actual session data length from the header and return it
for parsing.

Add verbose debug messages for the status report and the session data
content.

Testing Done: DB3.1: hotplug/unplug speaker module, check connections.

Change-Id: I50fb647efaa5d74fea4c7d9705014e5492c36f58
Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>