SW-3364: SVC: At boot, enable Switch port interrupts after interface power-on
Unipro port interrupts need to be enabled initially for the LinkUp and
Mailbox IRQs to be signalled by the Switch.
Later the port interrupts are enabled and disabled as part of
the hotplug/unplug sequence.

This fixes a regression on ES2 where the Switch port interrupt were
kept disabled at boot time.

Testing Done: DB3.1/DB3.5: Check connection state at boot on the SVC
('svc r') and the AP ('lsgb'); check modules hotplug

Change-Id: I45323b990d7e42fc80830255fbccbc6c94384ee5
Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>