SW-753: drivers: gpio: add gpio_is_valid()
Add a function that enables testing of whether a GPIO is valid. This is
necessary because simply checking that a GPIO is below the total line
count for the system does not account for the fact that GPIO chips can
have arbitrary bases, so there can be "holes" in the ID space.

Testing Done: tested in a later patch

Change-Id: Ia4b9c6f5dd2130737a3a2b85b0292db2b6bf1e59
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>