SW-3349: UniPro: Properly enable/disable the UniPro CSV feature
UniPro on the Toshiba bridge chips is implemented in two parts, multiple
CPort buffers and the UniPro stack. The CSV feature on a CPort buffer
must be configured the same as the CSV feature for the corresponding
CPort on the UniPro stack, otherwise operation is unpredictable.

When a Connection is created in the current code, CSV on the UniPro
stack is disabled while CSV on the CPort buffer is left in the default
state (enabled).

This patch disables CSV on the CPort buffer during UniPro init. When the
AP sends a request to enable a CPort, T_CPortFlags on the UniPro stack
for that CPort is read and CSV on the corresponding CPort buffer is set
accordingly.  When the AP sends a request to disable a CPort, CSV on the
corresponding CPort buffer is disabled.

Testing Done: Compile only. The infrastructure for dynamically
enabling/disabling a CPort is not currently available in Greybus.
Registers on the CPort buffer were manually written to ensure setting
the values as used in the patch work as expected.

Change-Id: I7af511b82c6e7ab4df3ce4c9ef75041fc03455ea
Signed-off-by: Rob Johnson <gtgprods@gmail.com>