SW-3061: UniPro: Add UniPro CSD flow control type
A UniPro port on the Toshiba bridge chips is implemented in two parts,
multiple CPort buffers and the UniPro stack. Both parts must be configured the
same to get the desired operation.

When a Connection is created in the current code, T_CportFlags on the
UniPro stack can enable any flow control mode, i.e., E2EFC or CSD, with
or without CSV. However, the CPort buffer is always enabled for E2EFC.

In this patch, when the AP sends a request to enable a CPort,
T_CPortFlags on the UniPro stack is read and the CPort buffer is set
accordingly, allowing for any of the flow control modes.

A special case is added for the Control CPort on a remote module, i.e.,
not APBA, to set T_CPortFlags in the initialization code since the
bridge in the remote module cannot talk to the AP before the Control
Connection is established.

Testing Done: Compile only. The infrastructure for dynamically
enabling/disabling a CPort is not currently available in Greybus.
Registers on the CPort buffer were manually written to ensure setting
the values as used in the patch work as expected.

Change-Id: I02a7c65e01387fee96c27db07f48f13d0d394805
Signed-off-by: Rob Johnson <gtgprods@gmail.com>