SW-3522: DWC OTG: Add a way to disable USB SG DMA enhancement
Currently, we can't disable SG DMA enhancement.
Add a way to do it at build time to help in USB debugging.
Disabling the enhancement can exonerate it when there is USB issue.
In addition, if this is the cause of issue, we can use the ifdef / endif
to locate easily the changes that may cause the issue.

Testing Done: Tested with gb_loopback when the option is enable and disabled

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>

Change-Id: Idb6775197a518af297d45556fa9b6363e3fdd82f