SW-1623/SW-3765: svc/timesync: Add SVC timesync functionality
This patch adds the SVC TimeSync logic in. We setup a cascading 16bit/16bit
master/slave timer to account for the fact that both TIM2 and TIM5 the 32
bit timers have already been used for other purposes. This timer when it
rolls over the upper 16 bit word, then increments the upper 32 bits of the
64 bit frame time.

We take the AP refclk as a parameter and setup a hardware timer to divide
our local timer down to the rate of the APs refclk. In practical terms the
AP is clocking at 19.2MHz and our local peripheral clock is 96MHz so, this
means we just pre-scale by 5 and the hardware timer is good to go in terms
of tracking the frame-time.

We accept and enable command from the AP which gives :
count:        The number of TIME_SYNC strobes to drive
frame_time:   The AP mandated time at the first TIME_SYNC strobe
strobe_delay: The delay in microseconds between each TIME_SYNC strobe
strobe_mask:  A bit-mask of SVC device IDs to strobe
              APB1 and APB2 corresponding to WD8A and WD8B are always
              strobed and will be or'd into the strobe_mask by the SVC
refclk:       The afore mentioned AP refclk at this revision of the phone
              this is the 19.2MHz uber-stable, temperature compensated
              PMIC clock

Once enabled the strobe timer will fire 'count' number of times and will
write all GPIOs simultaneously (a global strobe) of all indicated
downstream TIME_SYNC WDx slaves. The SVC will log the local frame time at
each strobe (using the AP mandated frame_time seed) and if TIMESYNC_DEBUG
has been enabled will kick the thread timesync_debugd_main to print out a
bunch of time-sync specific metrics plus the local frame-time every second.
If TIMESYNC_DEBUG is not set then SVC will simply continue to track the
frame-time.

@ 19.2MHz starting from a count of zero it will take approximately 30,000
years of continuous uptime to rollover a 64bit frame-time.

We support timesync_disable() and when called all timers will be stopped
and frame-time will no longer be tracked by the SVC. In order to restart
tracking of frame-time the AP must issue a new timesync_enable() command
globally to all participating entities in the ARA system.

Once the SVC is in the TIMESYNC_ACTIVE state it is possible for the AP to
send arbitrary ping requests to the SVC. In this case the SVC will send a
single strobe to an indicated bit-mask of device-ids. The SVC will return
the frame-time at the rising edge of the outbound strobe in the
response-phase of the greybus command. The AP can then subsequently poll
the various APBs and GPBs to find the frame-time at the event pulse on
each of those processors. This way we can verify the system time on each
processor - including the AP - for a given point in time.

Testing Done: Tested with AP/APB/GPB TIME_SYNC works well enough for FC.

Change-Id: Ia15a1cd191abea97f7e8655febf1c86b0b8344f3
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>