SW-3450: gdmac: Add a cport_test app to reset cport
This checkin adds a UniPro command option to allow us to
reset cport when UniPro Tx on a cport is hung. The purpose
of this command option is allowing us to test the Unipro
TX recovery routines when a cport hangs due to various reason.

Testing Done: Developer tested with GDMAC enabled UniPro TX
data transfer and gb_loopback_test causing the channel program
on abp2 to hang and issuing 'unipro cport_reset 12' to cause
UniPro to reset the cport 12 and terminate the hanging channel
program.

Change-Id: I4efecc63579171b6339c2361d1ea7b453d71796f
Signed-off-by: Kim-Ping Mui <mui_kim@projectara.com>