SW-2850: gdmac: Terminate Channel if active operation is deleted
This checkin added code to terminate GDMAC channel program
if an operation-in-process is deleted. This change is needed
to recover GDMAC channel and atabl setting in the case of
remote module is removed and causing GDMAC channel program
to hang waiting for transfer request.

Testing Done: Developer tested with patch in code review 4159
and 4154 on top of the latest code to generate GDMAC channel
program hanging condition and cport_reset was used to for the
the cport to be reset as well as trigger the recovery routine to
terminate the channel program and release the dma operation
descriptor.

Signed-off-by: Kim-Ping Mui <mui_kim@projectara.com>

Change-Id: Ia0ba4a44afc6542ab0665b68d4e03fc1ac4c9f82