SW-3479: apbridgea: audio: Initialize cport->direction
When apbridgea_audio_register_cport() allocates a cport structure,
it initializes all of the members except for 'direction'.  That
member is then OR'd potentially causing the cport structure to
not be freed by apbridgea_audio_unregister_cport() like it should.

Fix this be initializing the direction member to zero when it is
allocated.

Testing Done: Played audio on an EVT1.5 + EVT1 & EVT1.5 audio modules.
              Removed audio module and cport structure was freed.

Change-Id: I34b3852dcf564b65d032224da529b72bcc7a6279
Signed-off-by: Mark Greer <mgreer@animalcreek.com>