SW-3480: apbridgea: audio: Don't free in use ring buf entries
When transmitting audio, unipro_send_async() is used to send
Greybus audio data messages to the modules.  Once a message
has been sent and the ring buffer entry containing the audio
data is no longer needed by the UniPro subsystem, the
apbridgea_audio_unipro_tx_cb() callback is called.  The callback
will eventually return the ring buffer entry to the I2S driver
so it can be filled with data again.

This works fine except there is a race when shutting down audio
transmission.  It is possible for the ring buffer entry to be
freed in apbridgea_audio_rb_free() while the UniPro subsystem
is still using it.  To prevent this, use a semaphore in the
ring buffer entry's data header to indicate when it is safe
for apbridgea_audio_rb_free() to free it.  This only happens
when the APBRIDGEA_AUDIO_FLAG_TX_PREPARED flag is cleared
which means that audio transmission has stopped and related
resources are being freed.

Testing Done: Used key touch sounds to quickly set up and shut down
              audio connections and no in-use ring buffer entries
              were freed.  Before this patch, freeing in-use ring
              buffer entries happened frequently.

Change-Id: I98ea40e4d6bdb3716b9a742b4269b53dfcf3c8b9
Signed-off-by: Mark Greer <mgreer@animalcreek.com>