SW-3346 Set DMA for UniPro TX with 6 channels as default
Modify KConfig so that a freshly-created nuttX defconfig will select
DMA for UniPro TX, with 6 GDMAC channels allocated, by default.  This
default can be overridden as desired.

Due to SW-3533, DMA for UniPro TX is scoped to ES3 Bridge Silicon only.
Even if the bridge firmware is built with DMA for UniPro TX only, if ES3
silicon is not detected at runtime, the firmware will fallback to memcpy.

Testing Done: Developer-tested the SW-3346 patch series overlaid
on the SW-3533 patch series with Build #281 on a DB 3.2 (ES3 Silicon)
using gb_loopback-test transfer test with the current defconfigs on nuttx
master.  Synchronous tests, and asynchronous tests with a pipeline
depth of 3 and 1000 iterations each at packet sizes: 128, 256, 512,
1024, and 2020 completed without any obvious regression.

Change-Id: Id68ce5fd01e285874e5ed3a374fcbec708150f79
Signed-off-by: Mitchell Tasman <tasman@leaflabs.com>