SW-4895: svc: board: initialize board's switch regulator
In the board_init functions, we initialize the switch regulator for
each board.  Currently this is done by referencing the address of
a global switch regulator structure.  Instead, this function should
be initializing the switch regulator that's associated with the
board being initialized.

In all cases currently, the board info's switch regulator structure
is the same as the global being initialized now.  But initializing
the global hides the fact that the switch regulator is in fact
associated with the board info.

While we're at it, add error handling for the case that vreg_config
fails in db3_board_init().

Testing done:
Built resulting code.  Booted the built result.

Change-Id: I6f12c8a499a0928fad311b45ef60938a4700d253
Signed-off-by: Alex Elder <elder@linaro.org>