SW-3525: Nuttx RAMLOG: Add an option to overwrite the buffer
Currently, the RAMLOG return -EBUSY when it reach the end of buffer.
But usually, we want to overwrite the buffer to keep the end of logs.
Add an option in kconfig to overwrite the buffer.

Testing Done: Write a bunch of log in RAMLOG and check that logs are the last.

Change-Id: I34a1268d03d85afb8208d0bf4733a2b10f0f985d
Signed-off-by: Alexandre Bailon <abailon@baylibre.com>