ENG-673 interface: supply VSYS before ejecting
If the detect pin for an interface is asserted, but the interface is not
powered, apply power to the interface before ejecting. This enables the
ejection of dummy modules.

Testing Done: inserted two dummy modules in an EVT1.5 and used both the
'svc x' interface and the Ara key long press to eject the modules.

Change-Id: I4cc3c880de05f6a0324aaabcd370a00ac417f4fe
Signed-off-by: Jeffrey Carlyle <jcarlyle@google.com>