SW-3552: tsb: spi: simplify bitbang driver
It is not necessary anymore to request the pinshare on GPIOs. The GPIOs API
does it for you.

The driver is currently broken and doesn't build anymore but this change was
made in case someone decides it is worth it to fix this driver.

Testing Done: none, the driver is already broken and does not compile

Change-Id: Ic92fcdc723984698a5a1a818dffef2af45222b20
Signed-off-by: Fabien Parent <fparent@baylibre.com>