SW-3311: spi: remove unimplemented features (async and DMA)
The SPI API defines asynchronous and DMA transfers but they are not implemented
in the corresponding toshiba driver. It seems quite irrelevant to have
unimplemented features that need then to be documented in the FW documentation.

We can re-add them later if ever it becomes necessary (because of the greybus
spi protocol changes, or because the spi driver is updated).

Testing Done: compiled

Change-Id: I42c05009450f50f7f726bf642ad460603fe90324
Signed-off-by: Joel Porquet <porquet_joel@projectara.com>