SW-3311: spi: rename misnamed structure field "bpw"
The SPI master capabilities structure defines a field that is supposed to
contain the mask of the supported bits per word configurations. This field is
currently named "bpw" which is confusing since other structures actually define
"bpw" as being a value and not a mask.

Rename into "bpw_mask" to avoid confusion.

Testing Done: compiled

Change-Id: I9ea70d76037f811bafbcf133b7888140b81800d2
Signed-off-by: Joel Porquet <porquet_joel@projectara.com>