SW-3311: spi: rename data structure "xfer_curr_info" and internal fields
Rename the data structure "xfer_curr_info" into "xfer_info" because the aspect
"current" doesn't seem to be a part of the structure itself. It's only an
instance of this structure that is part of the SPI master device info that bears
the "current" aspect.

Also rename some of the fields of this structure. No need to duplicate the info
that it is indeed the current transfer (eg "xfer_curr->bpw_curr").

Testing Done: compiled

Change-Id: Ibe8d24a193f48d8d8cacf425a6675ebd4d412d49
Signed-off-by: Joel Porquet <porquet_joel@projectara.com>