SW-3311: spi: change the semantic of field "cs_high"
In the current implementation, cs_high is a bitfield included in the global info
structure of the spi driver. Each bit of cs_high contains whether the
corresponding spi slave is active on cs high (or cs low).

It is unnecessary to keep this information for each spi slave, because the
"cs_high" mode is given as part of each greybus transfer.

This patch removes the bitfield nature of this field and makes it a 0/1 value.

We still need to stored this information because the selection of the mode and
the activation of a specific spi slave are performed by two different function
calls.

Finally, the same way the bits-per-word parameter is specific to each transfer,
cs_high is also specific to each transfer. As such, it shouldn't be part of the
global data structure but part of the transfer data structure.

Testing Done: compiled

Change-Id: Idc78f409073f20c16d9fc1dcd843f9252be18422
Signed-off-by: Joel Porquet <porquet_joel@projectara.com>