SW-3311: spi: transform a driver parameter into a Kconfig option
In the current implementation, it's up to the module maker to provide a data
structure containing the parameter whether the SPI driver should use GPIOs for
chip selecting SPI slaves or the standard CS mechanism (via the CS pins).

The issue is that providing this parameters from the board custom code is not
really supported and forces an unnecessarily complex design to be put in place.
Instead such a parameter could be easily defined by a Kconfig option (like the
selection of the speed for the I2C driver) while offering the exact same
behavior.

Testing Done: compiled

Change-Id: Ia3b6bd0a5bc1bafa07d6e2bc55933e56337c2e2e
Signed-off-by: Joel Porquet <porquet_joel@projectara.com>