SW-3876: SVC PM: Eliminate possibility to enter PM_SLEEP
On ARA SVC the deepest allowed sleep mode of STM32 is Stop (=Nuttx
PM_STANDBY). The StandBy mode of STM32 is forbidden because it does not
allow the wake-up sources provided by the HW.
This patch targets STM32 Stop mode for both Nuttx PM_STANDBY and
PM_SLEEP modes.

Testing Done: Check idle mode of STM32; results to be added to JIRA SW-3279.

Change-Id: Iffcb563fc9b335e217b8c75c1f9e7425e70d91b3
Signed-off-by: Vladimir Mladenov <mladenov_vladimir@projectara.com>
Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>