SW-3311: spi: redesign SPI API
The SPI API and its implementation suffer from various flaws:

1/ the chosen abstraction level doesn't not apply to the device model in the
sense that a SPI board device was supposed to represent a group of SPI slaves
instead of only one slave.

2/ there are duplicated information in the declaration of SPI slaves. For
example, a SPI slave active on CS high has to specify it in its mode _and_
indicate the initial state of its GPIO CS output, whereas the latter can easily
be deduced from the first.

3/ the toshiba driver doesn't test the parameters against what slaves are
capable of doing: it is possible to specify a higher frequency than what
supports a slave and a different mode or bpw.

4/ the toshiba driver only uses the SPI board device for initialization and
copies everything locally.

5/ the toshiba driver only supports 2 SPI slaves even when using GPIO CS lines.

6/ the toshiba driver manages the GPIO lines for CS, even though this management
should belong to the SPI board interface.

This patch reworks the SPI API and solves the aforementioned issues. In
particular, it changes the abstraction of SPI board. Like in Linux, a SPI board
now represents one SPI slave. This way it's possible to use the device model at
its full potential: for example, using a GPIO resource to indicate on which GPIO
line the CS is attached to.

Improve the SPI board interface so that the toshiba driver can use it
directly instead of copying all the data structures locally. Also better
separate what each API provides so that the toshiba doesn't perform actions that
belong to the SPI board API.

Support for more than 2 SPI slaves when GPIO CS is selected. The SPI master
internally thinks there is only one SPI slave, and the manual CS management
enables a "unlimited" number of slaves.

Testing Done: compiled for me, but fully tested by Mike. Tested on the real
SPI tutorial with an accelerometer and well as with the oscilloscope

Change-Id: Id1389e7d6ad89bed6cd476dddf1534fa73860e2f
Signed-off-by: Joel Porquet <porquet_joel@projectara.com>