SW-3939: PWM: PWM silently fails if the freq & duty cycle are set too fast
The tsb_pwm.h file define the DIV value is 16, so the maximum of controller clock
is 48MHz/16 = 3MHz and then the internal lock timing is 333nS.
In this case, if the duty_cycle < 333nS, the current driver will set duty to "0" and
the activate wave has not work.
Therefore, we need to response failed if the duty_cycle < internal clock timing.

Testing Done: Verified on EVT1.5/GPB module and DB3.5/Red test module.

Change-Id: I0d4d477fe7380be3f4d38162ee682aec53704547
Signed-off-by: Philip Yang <yang_philip@projectara.com>