SW-3382: SVC: Do not busy wait during the WAKEOUT pulse
Assert the line and schedule the work to de-assert the line via a work
queue. This prevents the contention of the Nuttx HPWORK work queue.
The line is de-asserted and the scheduled work is canceled in case of
hot unplug or a power off of the interface.

Testing Done: DB3.5/EVT1.5: Check hotplug/unplug and WAKEOUT pulse
length

Change-Id: If86b0cafdd13cd81d4a3fabdc43d9e6c31ade740
Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>