SW-3252: svc: add kconfig option for evt2 phone with evt1.5 modules
Google requested that we merge the code for evt2 but we should for now keep
compatibility with evt1.5 modules since evt2 modules might be delayed and not
there during the bringup.
This commit adds a kconfig option to select whether we want to build a binary
with EVT2 module support or EVT1.5 module support when using a EVT2 phone.

Testing Done: none

Change-Id: I96821bd6960f2e8c07a62eaa7f10bcb0447bfc6b
Signed-off-by: Fabien Parent <fparent@baylibre.com>