SW-4137: tsb: aggregate the PM activity defines in a single file
Currently the values passed to pm_activity() are scattered across many
driver files in tsb. For now we just set them to the highest value
accepted by the pm framework, but at some point we will have to start
fine-tuning them.

Move the defines into tsb_pm.h so that we will not have to search for
them throughout the source code.

Testing Done: compiles fine

Change-Id: Ic379b7cd19123d8f668411f6e2c3d59bef02aff0
Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>