SW-3382: SVC: Add a nofification for LinkStartupInd from the bridges
When the bridges are booted and have Unipro initialized they initiate
LinkStartup requests to the Switch. In that case the LinkStartupInd IRQ
bit is set in the corresponding port TSB_InterruptStatus. On
successful LinkStartup the SVC end the WAKEOUT pulse and requests a manual
LinkStartup from the Switch to the bridge.

This is in line with the new hotplug sequence, where the Switch is
not the initiator of the LinkStartup requests.

This also fixes some issues with debug and error messages generation.

Testing Done: DB3.1, DB3.5, EVT1.5: check module detection at boot and on
hotplug/unplug.

Change-Id: Iaab51a6d4e9454bcd60de89fe0284c2c378d58ae
Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>