SW-3382: SVC: Implement bridge-driven LinkUp
The new hotplug mechanism is going in the direction of the Switch
not initiating the LinkUp request to the bridges. Instead the bridges
are continuously requesting LinkUp to the Switch.

On ES2, the Switch tries to LinkUp with bridges forever. Eventually
the bridges and Switch are successfully LinkingUp.

On ES3 since the Switch port and port IRQ is enabled before the bridges
boot, the LinkStartupInd from the bridges will trigger a notification
which in turn triggers a manual LinkStartup from the Switch to the
bridges.

This patch enables the Switch port and port IRQ before the WAKEOUT signal
generation and does not initiate the LinkStartup request from the
Switch.

While at it refactor the power_on function for proper error handling.

Testing Done: DB3.1, DB3.5, EVT1.5: check module detection at boot and
on hotplug/unplug.

Change-Id: I44a99a708e5060815928b14f8201723cf3469832
Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>