SW-4241: tsb: gpio: clear irq type before setting the register
Switching the irq type setting would not work and we
would be stuck with irq's in "both" directions even though
only rising or falling was selected.

This is beacause the type register was not cleared before
writing, and was only "or'd" with the new value, so clear
the last setting before writing a new type value.

Testing done: Through sysfs set the irq to rising and verify
that an irq is only trigger in transitions from low to hig,
do the same with falling.

Change-Id: I4d7f6b093718bb0a240b02bc55858372d1e14354
Signed-off-by: Axel Haslam <haslam_axel@projectara.com>