SW-3851: pwm: split interrupt state into two params for callback
Change the prototype of the PWM event callback to accept two parameters:
the mask of the PWM generators which have raised an event and the
corresponding errors for that event. Update the register_callback()
function to accept the same two parameters and a mode parameter for the
interrupt.

Testing Done: compiled and tested with module

Change-Id: I25bbb19c1ef8132201726c7e5bc50cebe0d3d810
Signed-off-by: Michael Mogenson <michael.mogenson@leaflabs.com>