SW-3382: SVC: Prevent the module ejection code to sleep with locks held
The module ejection code sleeps in the workqueue context with
the interface lock held. This breaks the Nuttx HPWORK work queue.
Split the ejection code in two halves, one for the module release
signal assertion and another that de-asserts the signal. The latter
is scheduled in the Nuttx work queue.

Note: the ejection of all modules need to happen serially in order to
keep the power consumption under control. The code introduced by this
patch triggers all ejections in parallel. A fix for that is in progress.

Testing Done: DB3.5, EVT1.5: Check hotplug and module ejection. Not
tested: ejection of all modules; EVT2

Change-Id: I53ff8c37fe810c49badea4c67958094eb1986936
Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>