SW-4323: irq: add private data to IRQ handlers
Add a void *priv parameter to the xcpt_t IRQ handler prototype. Modify
irq_attach() to accept a private data argument, store it in an array, and
return it to the attached IRQ handler when irq_dispatch() is called.

Update irq_unexpected_isr() to match the new xcpt_t type and update all
IRQ handlers and calls to irq_attach() in Nuttx to the new changes.

The purpose of these changes is so we can pass a pointer to a device
structure when attaching an IRQ handler and receive it back as a private
data argument when an interrupt occurs. The device driver can then
determine which device raised the interrupt and access any device info.
This makes device drivers truly generic by not having to save a global
pointer to the device structure.

Testing Done: built all

Change-Id: I88b1da5d8fae6ccd407b04eb366bdfad40aef61c
Signed-off-by: Michael Mogenson <michael.mogenson@leaflabs.com>