SW-1623: svc/timers: Add timer identifier to timer structure
On the STM32 some timers have features that other timers do not, including
some register and/or bit-fields depending on which timer block. In order to
ensure that the timer extensions we are adding do not accidentially get
used against an invalid timer block, we need to identify which timer is
which within the stm32_tim.c functions. This can be done by adding a
parameter to the various functions exported by stm32_tim.c or it can be
done by adding an internal field to the stm32_tim_dev_s structure.
This patch extends out that structure since it's a lower cost in terms of
code churn than adding a parameter to all of the stm32_tim.c functions
would be.

Change-Id: Ic06d4a8ad4d1debcdcc7fd72004422a371ab84f1
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>