SW-4419: Fix audio streaming crashed when inserting a speaker module
This checkin fixed GDMAC driver crashing issue when insert the
third speaker module into the device while Google Player is
streaming music. The root cause of this issue was that we were
using a sequence of if statements in tsb_dma_process_completed_op(),
in tsb_dma.c, to invoke callback function. If a callback released
the dma op we would end up with the next if statement trying to
access the already released dma op to determine the dma op's
state and mistakenly invoke the callback function again to
process the the already released dma op resulting with an highly
unpredictable system error.

Testing Done: Developer test on build #359 with Google player
playing music and insert/remove the third speaker for more than
30 times without causing any crashing or breaking audio
streaming.

Change-Id: I05592eb38d8d19beebb79b4d5c0b9656e1ff9cab
Signed-off-by: Kim-Ping Mui <mui_kim@projectara.com>