SW-4599: bridge_etm: remove pinshare saving
This application saves the previous state of the ETM pinshare bit before
clearing it. It doesn't seem necessary to do that since every driver that wants
to use the pinshare bits are responsible for setting up the correct value for
them.

It's not a big issue but I want to get rid of tsb_get_pinshare() and it's the
only occurrence.

Testing Done: compiled

Change-Id: I00a114d75d8e9da7563809de7627abf8f2aeac56
Signed-off-by: Joel Porquet <porquet_joel@projectara.com>