SW-4599: pinshare: introduce new pinshare subsystem
The biggest issue of the former pinshare API was that it wasn't possible
for two drivers to share pinshare bits.

For example, on GPBridge/ES3, it was impossible to activate the pins
GPIO3, PMW0 and SPIM_CS1_N as they all depend on PINSHARE[2] even though
they all want the same value for it, ie PINSHARE[2] = 0.

Instead of giving full access to the pinshare management to the drivers,
this new API provides a layer of abstraction that hides the gory
details. Now drivers request only the pins or groups of pins they want.
The pinshare subsystem is responsible for giving them access according
the internal status.

Before:

        /* configure PINSHARE to get ownership of UART_CTS and UART_RTS
         * pins */
        tsb_request_pinshare(TSB_PIN_UART_CTSRTS | TSB_PIN_GPIO9);
        tsb_clr_pinshare(TSB_PIN_GPIO9);
        tsb_set_pinshare(TSB_PIN_UART_CTSRTS);

After:

        /* get UART_CTS and UART_RTS pins */
        tsb_pin_request(PIN_UART_XTS);

This first patch only introduces the subsystem. Following patches will
adapt drivers one by one to facilitate reviews. The last patch will
remove the former API.

Testing Done: compiled

Change-Id: Icba4fac0482bdbcb25a85e935c29b4c676ec09d4
Signed-off-by: Joel Porquet <porquet_joel@projectara.com>