SW-4405: greybus-utils: cleanup manifest generation
In Kconfig, change the name of the options to make them more explicit.

In Makefile, remove some variables that are not necessary and reorganize a bit
the various cases.

This patch is in preparation of receiving a binary manifest file for
OOT_MANIFEST, instead of a source manifest that we have to compile. This way the
compilation will be done by the FDK and errors will be more easily detectable.

Testing Done: compiled

Change-Id: Ia53a17def9f9b9eae47a4020d19b41215b95b211
Signed-off-by: Joel Porquet <porquet_joel@projectara.com>