SW-1777: greybus-utils: Don't throw an error when cleaning
This only happens when doing in-tree development, which I'm not sure many people
do (but still, it's nice to fix long time bugs).

If the current .config defines CONFIG_OOT_MANIFEST, then the corresponding
Makefile (apps/greybus-utils/Makefile) expects the global variable OOT_MANIFEST
to be set. But when cleaning, it raises an error because in that case
OOT_MANIFEST is not set.

Testing Done: cleaned a tree that defines CONFIG_OOT_MANIFEST without an error

Change-Id: If6ad0a8fdc29364ed875f4ce00fe69c44637ca06
Signed-off-by: Joel Porquet <porquet_joel@projectara.com>