SW-3172: Break off the USB OTG DMA arbiter
The USB OTG DMA arbiter give priority to bulk out endpoints.
The test "gb_loopback_test -t transfer -i 10000 -s 128 -p -x -c 100 -o 1000"
breaks the "round robin" sheduling of in and out requests on the host.
Because of the priority of out endpoints, a transfer from BUFRAM to TxFIFO
for in endpoints may interrupted.
If too many out transfer are scheduled, at sone point,
the interrupted transfer will never resume even when there are not
out transfers to perform anymore.

To fix this issue, enable the global nak while performing an in transfer.
If an out transfer is in progress, it will complete it and then start to
NAK on the out endpoints. It will prevent the BUFRAM to TxFIFO to be
interrupted by a out transfer, and then prevent the issue to happen.
Once the transfer to FIFO complete, clear the global NAK.

Note:
The current version of this patch may let the global NAK
be cleared before the end of of all bulk in transfer.

Testing Done:
  Run gb_loopback_test -t transfer -i 10000 -s 128 -p -x -c 100 -o 1000
    Greybus doesn't restart because of SVC ping timeout
  Measured the performances while playing audio on speaker module
     Can't see any performance changes

Change-Id: I19c72081df14ed7ba27666625073f06a77cf908e
Signed-off-by: Alexandre Bailon <abailon@baylibre.com>