SW-4254: Remove all DMA ops from channel queue when channel is released
Existing DMA code removes the DMA ops from the channel queue
only if the ops are in the queued state and skips the ops that
is in the starting state or running state when the channel is
being released. This checkin added code to remove all pending
DMA ops, including the ops in the starting state or running
state, from the channel queue when the channel is being
released.

Testing Done: Developer test on build #359 with Google player
playing music and insert/remove the third speaker for more than
30 times without causing any crashing or breaking audio
streaming. The new code path was tested with developer's
private I2S DMA enabled driver to ensure I2S driver can
allocate, free DMA channel, and the ops on the channel queue
were removed correctly as AP starts and stops streaming
audio. gb_loopback_test was working as expected as well.

Change-Id: I871dd40a4c0f0aed053e9ad34170ffac0fba81ff
Signed-off-by: Kim-Ping Mui <mui_kim@projectara.com>